Update change_list.html
Was getting an error looking at the product list in the Cartridge admin.  It flagged {% date_hierarchy cl %} as the problem.

I was unable to figure out where the date_hierarchy tag comes from or what it does, but I did see in the code at https://github.com/sehmaschine/django-grappelli/blob/master/grappelli/templates/grp_doc/change_list.html
the above modifcation.  It checks to see if the variable exists before accessing it.  This fixes the error I was seeing locally, hopefully it will be useful generally.